fix(WPTestCase) move PHPUnit overridden methods to trait to allow dynamic definition
